# Improve siphash performance for longer data