@alexcrichton 2015-07-27
@bors: r+ 27c44ce9c3be36d49b829e3dfbdcc983bddd727d